Renamed package (bbe) to be shorter and ignore useless files
