+++ .gitignore
@@ -10,3 +10,9 @@ debug.log
 logs/
 
 target/
+
+*.class
+
+*.ctxt
+
+*.bluej
