--- bbegameboard/Coin.ctxt
@@ -1,13 +1,0 @@
-#BlueJ class context
-comment0.target=Coin
-comment1.params=
-comment1.target=Coin()
-comment2.params=x\ y
-comment2.target=Coin(int,\ int)
-comment3.params=
-comment3.target=int\ getX()
-comment4.params=
-comment4.target=int\ getY()
-comment5.params=
-comment5.target=java.lang.String\ getImagePath()
-numComments=6
