--- bbegameboard/EntityReflection.ctxt
@@ -1,45 +1,0 @@
-#BlueJ class context
-comment0.target=EntityReflection
-comment1.params=entity\ backupImgPath
-comment1.target=EntityReflection(java.lang.Object,\ java.lang.String)
-comment10.params=b
-comment10.target=void\ setDown(boolean)
-comment11.params=b
-comment11.target=void\ setLeft(boolean)
-comment12.params=b
-comment12.target=void\ setRight(boolean)
-comment13.params=b
-comment13.target=void\ setSpace(boolean)
-comment14.params=b
-comment14.target=void\ setEnter(boolean)
-comment15.params=b
-comment15.target=void\ setW(boolean)
-comment16.params=b
-comment16.target=void\ setA(boolean)
-comment17.params=b
-comment17.target=void\ setS(boolean)
-comment18.params=b
-comment18.target=void\ setD(boolean)
-comment19.params=
-comment19.target=void\ crashMuenze()
-comment2.params=
-comment2.target=java.lang.Object\ getEntity()
-comment20.params=
-comment20.target=void\ crashSpieler()
-comment21.params=o
-comment21.target=int\ toInt(java.lang.Object)
-comment3.params=
-comment3.target=Hitbox\ getHitbox()
-comment4.params=
-comment4.target=java.awt.Image\ getImage()
-comment5.params=
-comment5.target=int\ getX()
-comment6.params=
-comment6.target=int\ getY()
-comment7.params=
-comment7.target=java.lang.String\ getText()
-comment8.params=
-comment8.target=void\ run()
-comment9.params=b
-comment9.target=void\ setUp(boolean)
-numComments=22
