--- bbegameboard/Hitbox.ctxt
@@ -1,25 +1,0 @@
-#BlueJ class context
-comment0.target=Hitbox
-comment1.params=x\ y\ width\ height
-comment1.target=Hitbox(int,\ int,\ int,\ int)
-comment10.params=
-comment10.target=int[]\ getNormalizedMoveVector()
-comment11.params=other
-comment11.target=void\ determineIntersections(Hitbox)
-comment2.params=x\ y
-comment2.target=void\ setCoordinates(int,\ int)
-comment3.params=
-comment3.target=java.awt.Polygon\ getPolygon()
-comment4.params=other
-comment4.target=boolean\ collidesWithOther(Hitbox)
-comment5.params=other
-comment5.target=boolean\ collidesWithOther_fromAbove(Hitbox)
-comment6.params=other
-comment6.target=boolean\ collidesWithOther_fromBelow(Hitbox)
-comment7.params=other
-comment7.target=boolean\ collidesWithOther_fromLeft(Hitbox)
-comment8.params=other
-comment8.target=boolean\ collidesWithOther_fromRight(Hitbox)
-comment9.params=
-comment9.target=java.awt.Polygon\ lastStepPolygon()
-numComments=12
