--- bbegameboard/KeyboardListener.ctxt
@@ -1,13 +1,0 @@
-#BlueJ class context
-comment0.target=KeyboardListener
-comment1.params=keys
-comment1.target=KeyboardListener(int[])
-comment2.params=e
-comment2.target=void\ keyTyped(java.awt.event.KeyEvent)
-comment3.params=e
-comment3.target=void\ keyPressed(java.awt.event.KeyEvent)
-comment4.params=e
-comment4.target=void\ keyReleased(java.awt.event.KeyEvent)
-comment5.params=
-comment5.target=java.util.Map\ getPressedKeys()
-numComments=6
