--- bbegameboard/package.bluej
@@ -1,80 +1,0 @@
-#BlueJ package file
-dependency1.from=EntityReflection
-dependency1.to=Hitbox
-dependency1.type=UsesDependency
-dependency2.from=Gameboard
-dependency2.to=EntityReflection
-dependency2.type=UsesDependency
-dependency3.from=Gameboard
-dependency3.to=KeyboardListener
-dependency3.type=UsesDependency
-dependency4.from=Gameboard
-dependency4.to=Coin
-dependency4.type=UsesDependency
-dependency5.from=Gameboard
-dependency5.to=Fenster
-dependency5.type=UsesDependency
-dependency6.from=Fenster
-dependency6.to=Gameboard
-dependency6.type=UsesDependency
-objectbench.height=184
-objectbench.width=1621
-package.divider.horizontal=0.6
-package.divider.vertical=0.7976118500604595
-package.editor.height=747
-package.editor.width=1511
-package.editor.x=0
-package.editor.y=0
-package.frame.height=1042
-package.frame.width=1659
-package.numDependencies=6
-package.numTargets=6
-package.showExtends=true
-package.showUses=true
-readme.height=60
-readme.name=@README
-readme.width=48
-readme.x=10
-readme.y=10
-target1.height=70
-target1.name=EntityReflection
-target1.showInterface=false
-target1.type=ClassTarget
-target1.width=120
-target1.x=340
-target1.y=70
-target2.height=70
-target2.name=Coin
-target2.showInterface=false
-target2.type=ClassTarget
-target2.width=120
-target2.x=260
-target2.y=290
-target3.height=70
-target3.name=Gameboard
-target3.showInterface=false
-target3.type=ClassTarget
-target3.width=120
-target3.x=80
-target3.y=370
-target4.height=70
-target4.name=Hitbox
-target4.showInterface=false
-target4.type=ClassTarget
-target4.width=120
-target4.x=460
-target4.y=140
-target5.height=70
-target5.name=KeyboardListener
-target5.showInterface=false
-target5.type=ClassTarget
-target5.width=130
-target5.x=210
-target5.y=540
-target6.height=70
-target6.name=Fenster
-target6.showInterface=false
-target6.type=ClassTarget
-target6.width=120
-target6.x=220
-target6.y=450
