--- package.bluej
@@ -1,31 +1,0 @@
-#BlueJ package file
-editor.fx.0.height=0
-editor.fx.0.width=0
-editor.fx.0.x=0
-editor.fx.0.y=0
-objectbench.height=345
-objectbench.width=1621
-package.divider.horizontal=0.6
-package.divider.vertical=0.6073248407643312
-package.editor.height=538
-package.editor.width=1511
-package.editor.x=0
-package.editor.y=0
-package.frame.height=994
-package.frame.width=1659
-package.numDependencies=0
-package.numTargets=1
-package.showExtends=true
-package.showUses=true
-project.charset=windows-1252
-readme.height=60
-readme.name=@README
-readme.width=48
-readme.x=10
-readme.y=10
-target1.height=82
-target1.name=bbegameboard
-target1.type=PackageTarget
-target1.width=120
-target1.x=70
-target1.y=10
